+++ src/modules/common/index.ts
@@ -1,7 +1,9 @@
-import { PointCloud } from '../point-types';
+import { PointCloud } from './PointCloud';
 import * as Utils from './utils';
 
 export default {
   PointCloud,
   ...Utils,
 };
+
+export * from '../common/point-types';
