+++ tests/readPCDHeader.test.ts
@@ -6,6 +6,7 @@ describe('readPCDHeader.test', () => {
     const pcl = global.pcl as PCL.PCLInstance;
 
     const filename = 'ism_test_cat.pcd';
+
     writeFile(filename, pcl);
     const header = pcl.io.readPCDHeader(filename);
     expect(header?.fields).toEqual(['x', 'y', 'z']);
